create a separate price table component to manage price and implement search and filter functionality
