+++ src/app/components/banner/BannerTable.tsx
@@ -26,7 +26,7 @@ const BannerTable: React.FC<BannerTableProps> = ({
 }) => {
   const { deleteBanner } = useDeleteBanner();
   const [currentPage, setCurrentPage] = useState(1);
-  const [perPage, setPerPage] = useState<number>(5);
+  const [perPage, setPerPage] = useState<number>(10);
   const [searchParams, setSearchParams] = useSearchParams();
   const [sortColumn, setSortColumn] = useState<string | null>(null);
   const [sortOrder, setSortOrder] = useState<"ASC" | "DESC" | null>(null);
@@ -126,16 +126,11 @@ const BannerTable: React.FC<BannerTableProps> = ({
 
   const handleSort = (column: string) => {
     if (sortColumn === column) {
-      if (sortOrder === "ASC") {
-        setSortOrder("DESC");
-      } else {
-        setSortOrder("ASC");
-      }
+      sortOrder === "ASC" ? setSortOrder("DESC") : setSortOrder("ASC")
     } else {
       setSortColumn(column);
       setSortOrder("ASC");
     }
-    fetchBanners();
   };
 
   const handlePageChange = (newPage: number) => {
@@ -165,7 +160,6 @@ const BannerTable: React.FC<BannerTableProps> = ({
             value={perPage}
             onChange={handlePerPageChange}
           >
-            <option value={5}>5</option>
             <option value={10}>10</option>
             <option value={20}>20</option>
           </select>
@@ -271,7 +265,6 @@ const BannerTable: React.FC<BannerTableProps> = ({
                         <td>{banner.banner_id}</td>
                         <td>
                           <img
-                            alt={banner.title}
                             className="rounded-lg size-20 shrink-0"
                             src={banner.image}
                           />
