+++ src/app/components/category/CategoryTable.tsx
@@ -36,7 +36,7 @@ const CategoryTable: React.FC<CategoryTableProps> = ({
   const [originalCategoryName, setOriginalCategoryName] = useState<string>("");
 
   const [currentPage, setCurrentPage] = useState(1);
-  const [perPage, setPerPage] = useState<number>(5);
+  const [perPage, setPerPage] = useState<number>(10);
 
   const [searchParams, setSearchParams] = useSearchParams();
   const [sortColumn, setSortColumn] = useState<string | null>(null);
@@ -174,16 +174,11 @@ const CategoryTable: React.FC<CategoryTableProps> = ({
 
   const handleSort = (column: string) => {
     if (sortColumn === column) {
-      if (sortOrder === "ASC") {
-        setSortOrder("DESC");
-      } else {
-        setSortOrder("ASC");
-      }
+      sortOrder === "ASC" ? setSortOrder("DESC") : setSortOrder("ASC")
     } else {
       setSortColumn(column);
       setSortOrder("ASC");
     }
-    fetchCategories();
   };
 
   const handlePageChange = (newPage: number) => {
@@ -214,7 +209,6 @@ const CategoryTable: React.FC<CategoryTableProps> = ({
             value={perPage}
             onChange={handlePerPageChange}
           >
-            <option value={5}>5</option>
             <option value={10}>10</option>
             <option value={20}>20</option>
           </select>
