+++ src/app/components/company/CompanyTable.tsx
@@ -17,7 +17,7 @@ interface CompanyTableProps {
 
 const CompanyTable: React.FC<CompanyTableProps> = ({ search }) => {
   const [currentPage, setCurrentPage] = useState(1);
-  const [perPage, setPerPage] = useState<number>(5);
+  const [perPage, setPerPage] = useState<number>(10);
   const [searchParams, setSearchParams] = useSearchParams();
   const [sortColumn, setSortColumn] = useState<string | null>(null);
   const [sortOrder, setSortOrder] = useState<"ASC" | "DESC" | null>(null);
@@ -116,16 +116,11 @@ const CompanyTable: React.FC<CompanyTableProps> = ({ search }) => {
 
   const handleSort = (column: string) => {
     if (sortColumn === column) {
-      if (sortOrder === "ASC") {
-        setSortOrder("DESC");
-      } else {
-        setSortOrder("ASC");
-      }
+      sortOrder === "ASC" ? setSortOrder("DESC") : setSortOrder("ASC")
     } else {
       setSortColumn(column);
       setSortOrder("ASC");
     }
-    fetchCompanies();
   };
 
   const handlePageChange = (newPage: number) => {
@@ -155,7 +150,6 @@ const CompanyTable: React.FC<CompanyTableProps> = ({ search }) => {
             value={perPage}
             onChange={handlePerPageChange}
           >
-            <option value={5}>5</option>
             <option value={10}>10</option>
             <option value={20}>20</option>
           </select>
