+++ src/app/components/coupon/CouponTable.tsx
@@ -15,7 +15,7 @@ const CouponTable: React.FC<CouponTableProps> = ({
     search
 }) => {
   const [currentPage, setCurrentPage] = useState(1);
-  const [perPage, setPerPage] = useState<number>(5);
+  const [perPage, setPerPage] = useState<number>(10);
   const [searchParams, setSearchParams] = useSearchParams();
   const [sortColumn, setSortColumn] = useState<string | null>(null);
   const [sortOrder, setSortOrder] = useState<"ASC" | "DESC" | null>(null);
@@ -109,16 +109,11 @@ const CouponTable: React.FC<CouponTableProps> = ({
 
   const handleSort = (column: string) => {
     if (sortColumn === column) {
-      if (sortOrder === "ASC") {
-        setSortOrder("DESC");
-      } else {
-        setSortOrder("ASC");
-      }
+      sortOrder === "ASC" ? setSortOrder("DESC") : setSortOrder("ASC")
     } else {
       setSortColumn(column);
       setSortOrder("ASC");
     }
-    fetchCoupons();
   };
 
   const handlePageChange = (newPage: number) => {
@@ -148,7 +143,6 @@ const CouponTable: React.FC<CouponTableProps> = ({
             value={perPage}
             onChange={handlePerPageChange}
           >
-            <option value={5}>5</option>
             <option value={10}>10</option>
             <option value={20}>20</option>
           </select>
