+++ src/app/components/product/ProductTable.tsx
@@ -25,7 +25,7 @@ const ProductTable: React.FC<ProductTableProps> = ({
   setIsSubmit,
 }) => {
   const [currentPage, setCurrentPage] = useState(1);
-  const [perPage, setPerPage] = useState<number>(5);
+  const [perPage, setPerPage] = useState<number>(10);
   const [searchParams, setSearchParams] = useSearchParams();
   const [sortColumn, setSortColumn] = useState<string | null>(null);
   const [sortOrder, setSortOrder] = useState<"ASC" | "DESC" | null>(null);
@@ -128,16 +128,11 @@ const ProductTable: React.FC<ProductTableProps> = ({
 
   const handleSort = (column: string) => {
     if (sortColumn === column) {
-      if (sortOrder === "ASC") {
-        setSortOrder("DESC");
-      } else {
-        setSortOrder("ASC");
-      }
+      sortOrder === "ASC" ? setSortOrder("DESC") : setSortOrder("ASC")
     } else {
       setSortColumn(column);
       setSortOrder("ASC");
     }
-    fetchProduct();
   };
 
   const handlePageChange = (newPage: number) => {
@@ -168,7 +163,6 @@ const ProductTable: React.FC<ProductTableProps> = ({
             value={perPage}
             onChange={handlePerPageChange}
           >
-            <option value={5}>5</option>
             <option value={10}>10</option>
             <option value={20}>20</option>
           </select>
