+++ src/app/hooks/banner/useGetBanner.ts
@@ -13,7 +13,7 @@ interface Banner {
 
 const useGetBanner = (
   pageNumber: number = 1,
-  perPage: number = 5,
+  perPage: number = 10,
   search: string = "",
   sortColumn?: string,
   sortOrder?: string
