+++ src/app/hooks/categories/useGetCategories.ts
@@ -11,7 +11,7 @@ interface Category {
 
 const useGetCategories = (
   pageNumber: number = 1, 
-  perPage: number = 5,
+  perPage: number = 10,
   search: string = '', 
   sortColumn?: string, 
   sortOrder?: string
