+++ src/app/hooks/company/useGetCompany.ts
@@ -28,7 +28,7 @@ interface Company {
 
 const useGetCompany = (
   pageNumber: number = 1,
-  perPage: number = 5,
+  perPage: number = 10,
   search: string = '',
   sortColumn?: string,
   sortOrder?: string
