+++ src/app/hooks/coupon/useFetchCoupon.tsx
@@ -21,7 +21,7 @@ interface Coupon {
 
 const useFetchCoupons = (
   pageNumber: number = 1,
-  perPage: number = 5,
+  perPage: number = 10,
   search: string = '',
   sortColumn?: string,
   sortOrder?: string
