+++ src/app/hooks/services/useGetServices.ts
@@ -11,7 +11,7 @@ interface Service {
 
 const useGetServices = (
   pageNumber: number = 1,
-  perPage: number = 5,
+  perPage: number = 10,
   search: string = "",
   sortColumn?: string,
   sortOrder?: string) => {
